requirements.txt
asyncpg==0.22.0
fastapi==0.85.1
passlib==1.7.4
pyjwt==1.7.1
python-dotenv==0.21.0
SQLAlchemy>=1.4.41
pydantic==1.10.2
uvicorn==0.17.6
typing-extensions==4.4.0
python-multipart==0.0.5
psycopg2-binary==2.9.3
Jinja2==3.1.2
cryptography==38.0.1
bcrypt==4.0.1
